+++ app/src/main/java/com/candybar/dev/applications/CandyBar.java
@@ -4,6 +4,8 @@ import androidx.annotation.NonNull;
 
 import candybar.lib.applications.CandyBarApplication;
 
+//import com.onesignal.OneSignal
+
 public class CandyBar extends CandyBarApplication {
 
     // Remove '/*' and '*/' to Enable OneSignal